+++ delete_pic.php
@@ -2,9 +2,11 @@
 session_start(); // Should we also delete it from the folder
 require ('./connection.php');
 require ('./valid_session_check.php');
-if (!empty($_POST['deletepath'])) 
+if (!empty($_POST['deletepath']) ) 
 {
+	
 	$path = trim($_POST['deletepath']);
-	unlink($path);
+	if ($path != "images/error.png")
+		unlink($path);
 }
 ?>
